admin_portal/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-filter==24.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
